requirements.txt
pyodbc==5.1.0
python-dotenv==1.0.0
fastapi
jinja2
azure-identity
uvicorn[standard]